nature+jardin-ile-de-la-motte/groupe-6/requirements.txt
pygame
joycon-python
hidapi
pyglm
pyserial
